reconfigured version bumper
